Fix to single quotes in signature verification. (#27)
* Fix to 'Quotes are included in the Arraybuffer'

Quotes being included in the ArrayBuffer is an unexpected behavior that creates a problem when other systems attempt to verify signatures created by easy-web-crypto. This fix allows strings passed in to not have the Quotes, while not changing the behavior of objects being passed in.

* Fixed a bug in sign and verify related to strings

Sign and Verify include opening and closing single quotes when a string is passed instead of an object. The single quotes are unexpected. This change fixes that issue. Possible breaking change for some.

* Added tests for data string

Added two tests for the data string in signature sign/verify